+++ app/test/shared/versions_test.dart
@@ -31,7 +31,7 @@ void main() {
     // This test is a reminder that if pana, the SDK or any of the above
     // versions change, we should also adjust the [runtimeVersion]. Before
     // updating the hash value, double-check if it is being updated.
-    expect(hash, 267686003);
+    expect(hash, 701498168);
   });
 
   test('runtime version should be (somewhat) lexicographically ordered', () {
@@ -74,9 +74,6 @@ void main() {
   });
 
   test('Flutter is using a version from the stable channel.', () async {
-    // Special exception for a release candidate
-    // TODO: remove once this becomes a stable release.
-    if (flutterVersion == '1.12.13+hotfix.4') return;
     final flutterArchive = await fetchFlutterArchive();
     expect(
         flutterArchive.releases.any(